+++ base/deprecated.jl
@@ -1360,6 +1360,9 @@ end
 # issue #25928
 @deprecate wait(t::Task) fetch(t)
 
+# Issue 25786
+@deprecate_binding DevNull DEVNULL
+
 # END 0.7 deprecations
 
 # BEGIN 1.0 deprecations