+++ base/docs/basedocs.jl
@@ -790,16 +790,16 @@ A variable referring to the last computed value, automatically set at the intera
 kw"ans"
 
 """
-    DevNull
+    DEVNULL
 
 Used in a stream redirect to discard all data written to it. Essentially equivalent to
 /dev/null on Unix or NUL on Windows. Usage:
 
 ```julia
-run(pipeline(`cat test.txt`, DevNull))
+run(pipeline(`cat test.txt`, DEVNULL))
 ```
 """
-DevNull
+DEVNULL
 
 # doc strings for code in boot.jl and built-ins